Fixed Xcode 14.2 warnings

+++ src/audio/coreaudio/SDL_coreaudio.m
@@ -1214,7 +1214,7 @@ static int COREAUDIO_GetDefaultAudioInfo(char **name, SDL_AudioSpec *spec, int i
             /* Some devices have whitespace at the end...trim it. */
             while ((len > 0) && (devname[len - 1] == ' ')) {
                 len--;
-                usable = len;
+                usable = (int)len;
             }
         }
